requirements.txt
streamlit==1.28.1
pandas==2.1.3
numpy==1.25.2
boto3==1.29.165
duckdb==0.9.1
pyyaml==6.0.1
openpyxl==3.1.2
xlrd==2.0.1 